+++ x-pack/plugin/eql/qa/common/build.gradle
@@ -5,6 +5,7 @@ dependencies {
   api project(':test:framework')
   api project(path: xpackModule('core'), configuration: 'default')
   api project(path: xpackModule('core'), configuration: 'testArtifacts')
+  api project(path: xpackModule('ql'), configuration: 'testArtifacts')
 
   // TOML parser for EqlActionIT tests
   api 'io.ous:jtoml:2.0.0'
